tilaussivu

+++ PizzeriaProjekti/PizzeriaProjekti/Models/LisatayteModel.cs
@@ -4,6 +4,6 @@
     {
         public int Lisatayte_ID { get; set; }
         public string LT_Nimi { get; set; }
-        public float LT_Hinta { get; set; }
+        public decimal LT_Hinta { get; set; }
     }
 }
